Removed previous test case file as it was not using React testing library
