+++ src/Breadcrumbs/Breadcrumbs.test.js
@@ -1,77 +1,81 @@
 import React from "react";
-import Enzyme, { mount } from "enzyme";
-import Adapter from "@wojtekmaj/enzyme-adapter-react-17";
+import { render, fireEvent } from "@testing-library/react";
 import Breadcrumbs from "./Breadcrumbs";
 
-Enzyme.configure({
-  adapter: new Adapter(),
-});
-
-describe("Breadcrumbs", () => {
-  it("should render one breadcrumb", () => {
+describe("Breadcrumbs tests", () => {
+  test("Dom tree of one breadcrumb should be returned when one breadcrumb is given as input ", () => {
     const title = "A title";
     const onClick = jest.fn();
-    const breadcrumbs = [{ title, onClick }];
+    const islastCrumb = true;
 
     const style = { background: "red" };
-    const className = "A className";
+    const className = "TestClassName";
 
-    const wrapper = mount(
+    const breadcrumbs = [{ title, onClick, islastCrumb }];
+    const { getByText, container } = render(
       <Breadcrumbs
-        style={style}
-        className={className}
         breadcrumbs={breadcrumbs}
+        className={className}
+        style={style}
       />
     );
+    const breadcrumbElement = getByText("A title");
 
-    expect(wrapper.find(".common__breadcrumbs").prop("style")).toEqual(
-      expect.objectContaining(style)
-    );
-    expect(wrapper.find(".common__breadcrumbs").prop("className")).toContain(
-      className
-    );
-
-    expect(wrapper.find(".breadcrumb__item").text()).toEqual(title);
-
-    // If only 1 breadcrumb, not clickable (current page open)
-    wrapper.find(".breadcrumb__item").simulate("click");
-    expect(onClick).not.toHaveBeenCalled();
+    expect(container.firstChild.className).toMatch(/common__breadcrumbs/);
+    expect(container.firstChild.className).toMatch(/TestClassName/);
+    expect(container.firstChild.style.background).toMatch(/red/);
+    expect(breadcrumbElement.textContent).toMatch("A title");
+    expect(breadcrumbElement.className).toMatch(/breadcrumb__item/);
   });
 
-  it("should render multiple breadcrumbs", () => {
-    const firstTitle = "A title";
-    const firstOnClick = jest.fn();
+  test("Dom tree of more than one breadcrumb should be returned when more than one breadcrumb is given as input", () => {
+    const firstBreadcrumbFunction = jest.fn();
+    const secondBreadcrumbFunction = jest.fn();
 
-    const secondTitle = "A title too";
-    const secondOnClick = jest.fn();
+    const style = { background: "red" };
+    const className = "TestClassName";
     const breadcrumbs = [
-      { title: firstTitle, onClick: firstOnClick },
-      { title: secondTitle, onClick: secondOnClick },
+      {
+        key: "1",
+        title: "test",
+        onClick: { firstBreadcrumbFunction },
+        isLastBreadcrumb: false,
+      },
+      {
+        key: "2",
+        title: "test2",
+        onClick: { secondBreadcrumbFunction },
+        isLastBreadcrumb: true,
+      },
     ];
+    const { getByText, container } = render(
+      <Breadcrumbs
+        breadcrumbs={breadcrumbs}
+        className={className}
+        style={style}
+      />
+    );
+    const firstBreadcrumbElement = getByText("test");
+    const secondBreadcrumbElement = getByText("test2");
 
-    const wrapper = mount(<Breadcrumbs breadcrumbs={breadcrumbs} />);
+    expect(container.firstChild.className).toMatch(/common__breadcrumbs/);
+    expect(container.firstChild.className).toMatch(/TestClassName/);
+    expect(container.firstChild.style.background).toMatch(/red/);
+    expect(container.firstChild.className).toMatch(/TestClassName/);
+    expect(container.firstChild.style.background).toMatch(/red/);
 
-    const firstBreadcrumb = wrapper.find(".breadcrumb__link");
-
-    expect(firstBreadcrumb.text()).toEqual(firstTitle);
-
-    firstBreadcrumb.simulate("click");
-    expect(firstOnClick).toHaveBeenCalled();
-
-    const secondBreadcrumb = wrapper.find(".breadcrumb__item").at(1);
-
-    expect(secondBreadcrumb.text()).toEqual(secondTitle);
-
-    secondBreadcrumb.simulate("click");
-    expect(secondOnClick).not.toHaveBeenCalled();
-    expect(firstOnClick).toHaveBeenCalledTimes(1);
+    expect(firstBreadcrumbElement.parentElement.className).toMatch(
+      "breadcrumb__item"
+    );
+    expect(firstBreadcrumbElement.textContent).toMatch("test");
+    expect(firstBreadcrumbElement.className).toMatch("breadcrumb__link");
+    expect(secondBreadcrumbElement.className).toMatch("breadcrumb__item");
+    expect(secondBreadcrumbElement.textContent).toMatch("test2");
   });
 
-  it("should render no breadcrumbs", () => {
-    const breadcrumbs = [];
-
-    const wrapper = mount(<Breadcrumbs breadcrumbs={breadcrumbs} />);
-
-    expect(wrapper.find(".breadcrumb__item").exists()).toEqual(false);
+  test("Empty div tag is returned when no breadcrumb is given as an input", () => {
+    const breadCrumbs = [];
+    const { container } = render(<Breadcrumbs breadcrumbs={breadCrumbs} />);
+    expect(container.firstChild).toBeFalsy();
   });
 });
